hexfile: shrink memory image buffer
We're using a 50Mb memory buffer raw data that we're about to upload to
the stm8 or that we've just downloaded from the stm8. If stm8gal runs on
an embedded system with limited ram, this huge buffer is a problem.

The data in this buffer is transferred 1:1 to/from the stm8. It cannot be
larger than the maximum flash size of the stm8, i.e. 256Kb.

Let's reduce the buffer size from 50Mb to 5Mb for now.