main: hexfile: avoid huge file buffer
The code allocates a buffer of LENFILEBUF bytes (50 Mb) to store a disk
file before it's converted and uploaded. This is a problem on embedded
systems where RAM is limited.

We can avoid this large buffer simply by allocating the buffer in
load_file, not in main. The load_file function reads the file size
anyway, it can allocate a buffer that's exactly the size of our file.

Now that we fill the entire buffer with our file, there's no need to
initialise it with 0x00 bytes.